refactor: move server/{debug,metrics} under server/api
